// src/app/api/model/consent-proof-dto-in.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ConsentProofDtoIn { 
    checkboxId?: string;
    documentHash?: string;
    eventTrusted?: boolean;
    screenX?: number;
    screenY?: number;
    timestamp?: number;
}