// src/app/api/model/legal-document-dto-out.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LegalDocumentType } from './legal-document-type';


export interface LegalDocumentDtoOut { 
    contentHash?: string;
    downloadUrl?: string;
    effectiveFrom?: string;
    /**
     * LegalDocumentType enum name
     */
    type?: LegalDocumentType;
    version?: number;
}